+++ changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2021-08-03
+### Changed
+- `system_message` -> `systemMessage`.
+
+### Added
+- `userJoined` event.
+- `userLeft` event.
+
 ## [1.1.1] - 2021-08-02
 ### Fixed
 - No longer emits `system_message` event concerning self.